Restructured functions.
