+++ src/apkcontroller/org_torproject_android/V16_6_3_RC_1/screen_6.py
@@ -8,13 +8,13 @@ import networkx as nx
 from typeguard import typechecked
 from uiautomator import AutomatorDevice
 
-from src.apkcontroller.script_helper import get_expected_screen_nrs
-from src.apkcontroller.Screen import Screen
-from src.apkcontroller.script_helper import (
-    get_screen_as_dict,
+from src.apkcontroller.helper import get_screen_as_dict
+from src.apkcontroller.org_torproject_android.V16_6_3_RC_1.app_specific_helper import (
     get_torified_item_index_dict,
     orbot_torifying_app_is_checked,
 )
+from src.apkcontroller.Screen import Screen
+from src.apkcontroller.script_orientation import get_expected_screen_nrs
 
 
 @typechecked
