+++ src/apkcontroller/org_torproject_android/V16_6_3_RC_1/screen_flow.py
@@ -1,5 +1,4 @@
 """Stores the flow logic of the script in a networkx graph."""
-from typing import List
 
 import networkx as nx
 from typeguard import typechecked
